+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add ip and timezone into email footer.
+- Add `TimezoneMiddleware` to set `display_timezone` session by ip.
+
 ## [2.14.2] - 2025-10-10
 
 ### Fixed
@@ -801,6 +808,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [2.14.2]: https://https://github.com/internetguru/laravel-common/compare/v2.14.1...v2.14.2
 [2.14.1]: https://https://github.com/internetguru/laravel-common/compare/v2.14.0...v2.14.1
 [2.14.0]: https://https://github.com/internetguru/laravel-common/compare/v2.13.0...v2.14.0